refactor: Simplified traverse
